ZIM 10.8.0 - Wrapper, Selector, darken, lighten
WRAPPER
https://zimjs.com/ten/wrapper.html
Added to Control Module after Tile()
A Wrapper wraps its content to the next row at a given width
The Wrapper is similar to the CSS Flexbox with the following:
ALIGNMENTS:
It has left, center/middle right then top, center/middle bottom
alignments
for whole rows and columns and then for inside the rows and columns.
These may or may not be active depending on the wrapper type below
WRAPPER TYPES:
The types are spacing (default), spread, stretch, or column
A colSize is available for the column type
A rowSize is available that works with any of the above types
SETTINGS:
There are also flip, reverse, bottomFull and col/row void settings.
MARGINS:
ZIM display objects can have margins left, top, right, bottom
that will shift objects relatively inside the Wrapper
without affecting other objects in the wrapper
WRAPPER WITH WINDOW AND LAYOUT
The Wrapper is resized automatically when added to a ZIM Window.
Set the Window resizeHandle parameter to true to let the user resize.
The Wrapper resizes and scales as a region in the ZIM Layout class.
The Window also resizes - so a Wrapper can go in a Window in a Layout
object.

SELECTOR
https://zimjs.com/ten/wrapper.html
Added to Display Module after Slider
A Selector acts on a ZIM Tile to provide an interactive selector that
highlights a tile element.
The selector works as a select bar and a select pad.

Selector has a selectedIndex parameter so we
adjusted all components with selectedIndex to include selectedIndex as a
parameter
BREAK - these were added just before the style parameter near the end of
the parameters for
RadioButtons, Indicator, List, Stepper, Dial (currentValue), Tabs, Pad,
Radial, ColorPicker
Now an initial selectedIndex property can be done in the parameters

WINDOW - IMPROVEMENT
https://zimjs.com/ten/window.html
added fullSize, fullSizeColor and resizeHandle parameters to Window
these work to maximize / unmaximize the window when a titleBar is
present
the resizeHandle lets the user resize the window from the bottom right
corner
Added support for automatic resizing of Wrapper

ANIMATE
Added pauseOnBlur parameter defaults to true
this pauses the animation when the window reduces or a different tab is
selected
and helps keep animations in sync - as ZIM uses a requestAnimationFrame
which slows down if the window is not showing
Also added the parameter to timeout and interval - although it defaults
to false there
added loopPick parameter after loopWaitParams BREAKs parameter order for
anything from rewind on
this will call ZIM VEE values for animate properties again at each loop
- thanks GSAP for the prompting
we always wanted to do this but CreateJS had no way to change the
properties once we start the animation
so we adjusted the ZIM version of CreateJS 1.2.4 to expose the animation
properties and use this to provide the feature
We will put in a pull request to CreateJS to have a line added.
After this if statement finishes: if (inject = this._injected) {
then add this.step = step; to expose the properties to ZIM animate.
Currently does not work with from set on a sequence.
Added a rate parameter near the end of animate() to change the rate of
speed of the animation
This calls to CreateJS timeScale which is more efficient than the
existing percentSpeed parameter
and can use ZIM VEE (Pick) to adjust each loop with loopPick set to true
Fixed sequence call to call only on last animation of a sequence (or a
loop of a sequence)
Set animate to css true if there is no Frame - that lets animate be used
without a Frame
Added example to dot animation to docs - also added description to props
parameter
For example: animate(new THREEJS.mesh(etc.), {"rotation.y":360*RAD},
5000);

BITMAP - BLITTING
Adjusted Bitmap to receive a DisplayObject and automatically use its
cache canvas
For example if circles is a Container of Circles. The old way:
var splat = new Bitmap(circles.cache().cacheCanvas).addTo();
circles.uncache();
becomes the following - if circles was cached it stays cached if not
then it stays not cached
splat = new Bitmap(circles).addTo();

COLORS
added color.lighten(ratio) and color.darken(ratio) methods to colors
as short-cuts to colorRange(color1, color2, ratio)
red.darken(.2) gives a slightly darker red
ratio is between 0 (red) and 1 (black)
these are methods of a String and also functions darken(color, ratio),
lighten(color, ratio)
Also adjusted colorRange to actually default to .5 ratio as the docs say

GENERAL
BREAK - added a percentClose parameter to Circle that defaults to true
set to false to not close the border on a Circle with a percent set
Thanks Yan Zhang for the suggestion!
Changed ZIM Game module Scorer to set type property to "Scorer" instead
of "Score"
All ZIM object should return the exact class name from which they were
made
In ZIM physics, for the contact callback function first parameter,
a border will have a type = "Border" and a side = "left", "right",
"top", "bottom"
but it is not really a ZIM Rectangle but just an object literal
placeholder
adjusted loc() to properly set at index when index is provided and it is
locating in same parent
moved move/mov style to after loc and pos so can move after positioning
in STYLE
Fixed Window swiper area adjust after scrolling with scrollbar - on
scrollBarDrag true
Adjusted interval to not use requestAnimationFrame - this better syncs
the interval with setInterval
thanks Racheli Golan for noticing time discrepancy with fast intervals
BREAK Changed the Stepper default stepperType to "number" rather than
"list"
Need to set the stepperType to list for a list of words for instance in
the stepper
Set the default max for the Stepper to 10 instead of 100