ZIM Cat 01 - patches - this is the NPM version of 01.
PATCHED
Adjusted the VERSION to cat/01/zim - thanks Stefan for the catch!
Adjusted stopAnimate(id, toEnd) to handle ids with toEnd - there was a
bug
and adjusted endTween() to work across multiple animations and added
suggestions to Docs. Thanks Disco for the report.
Removed a duplicate property deep in animate code and changed a var poo
to po
and updated diffX and diffY in Squiggle and Blob to be vars - thanks
Stefan for the finds.
Stefan Meyer from Uruguay ran ESLint on ZIM so we went in and updated a
bunch of issues like missing vars, duplicate definitions, semicolons,
etc.
Updated cloneAll() method to have style, group, inherit parameters
Fixed subsequent allowToggle settings in transform()
Adjusted show and hide of HotSpot() to use but rather than older backing
variable
Fixed noise wave in Synth() to run with oscillator.context
Fixed series({min, max}) to properly loop on max number
Adjusted series({min, max, step}) to create an array and then just use
array code - that means step has been added as well.
Fixed array creation to handle step properly.
Adjusted zimify() to not zimify() and object that already has had
zimify() called on it - thanks David Dorin for the report!
Went back and did 10.9.0 (popular ZIM SHIM version) and cat 00 as well
as this 01 and beyond.
Added maxConnections parameter to Frame and loadAssets() thanks Disco
for the suggestion
This flows through to the CreateJS PreloadJS to the loadQueue
setMaxConnections() method
Adjusted Organizer remove(index) to actually remove the item at the
index rather than at the selectedIndex - thanks Disco for the report
and fixed a Window scrollX and scrollY setting issue due to an added
timeout earlier.
The setting of the scroll values was setting to right position but then
the 50ms timeout
set the content position back to where the scrollbars were at.
Adjusted Window to also drag after 300ms when dragBoundary is set -
after enabled or disabled
This related to conflicting with damping on scrollbars or something like
that...
Fixed removeAt() for List to properly adjust the selectedIndex
Adjusted transformControls dispose() to turn off events on handles, etc.
- thanks Ami Hanya for find
Fixed KeyBoard to work again with blank Labels - thanks Ami Hanya.
Added scrollEnabled property to Window (and therefore List) - thanks Ami
Hanya for request.
Adjusted Dial to work with mouseMoveOutside when set to true in Frame
Made Radial and RadialMenu style and inherit parameters pass through for
its center circle and labels
Adjusted RadialMenu to not start with selection when adding rings
Renamed internal level property to ringLevel now that DisplayObjects
have level property
Added scrambled event for Scrambler which dispatches when scrambling has
stopped
LabelOnPath was needing stage update on toggle
Called toggle(true) on path to set clicking off path to hide controls.
Added frame parameter to Sprite so can load assets from Frame other than
zimDefaultFrame
Fixed pause(false) on MotionController to work properly - was a
lastSpeed scope issue
Adjusted noDrag() to properly close drag when slide is true so next
drag() will work properly
This was affecting the toggle of enable for Window and List - important
if animating a list for instance.
Adjusted the addAt() and the removeAt() of List to not go to the top
each time.
We do not know why this was put in there - it may have been to prevent
some bug.
We tested and it seems to work fine just leaving the list where it is.
Apologies for the List work... it is quite complicated as it uses Tabs
in Window and has all sorts of thing going on.
Added frame parameter to Pen just before the style parameter to let Pen
be used on other frames
Parallax addLayer() removeLayer() step() immediate() pause() - returns
object for chaining
Added frame parameter to Generator just before the style parameter to
let Generator be used on other frames
Fixed scaling stutter in non-Retina full mode after 10.7.1 fix - did not
patch other versions as non-Retina is unusual

OBJECT
Added object() global function or zim.object() if zns is true (ZIM
namespace pre-load variable)
This lets you get any named DisplayObject
Added nam() short chainable method to set the name property - thanks
Disco for the idea
Or use the name property (also parent.getChildByName() still works for
Container children with names set)

DRAGGABLE
Added a draggable property to DisplayObjects that calls a basic drag()
or noDrag()
This allows drag state to be recorded - for instance using ZIMON.
Thanks Stefan Meyer for the suggestion - and I think we had an earliers
request for this too...
removed the readOnly draggable property from PEN - BREAK