ZIM Cat 02 - updates to Generator and Sprite
GENERATOR UPDATES
https://codepen.io/zimjs/pen/XWjrmoQ
Added a noise() function to the Generator that does ZIM Noise simplex
methods automatically choosing 1D-4D.
This returns a number from 0 to 1 (matching Processing/P5js) as opposed
to -1 to 1.
Added a seed parameter to Generator to set the Noise seed if desired.
Added resetup(), redraw() and restamp() methods to Generator
To call the setup, and start the draw and stamp functions again
This is very cool with stamp as we can animate through stamped pictures
to basically animate the Generator
-- beyond just doing the draw animation - see
https://codepen.io/zimjs/pen/XWjrmoQ
Added clear and reset methods to Generator
The clear() method will clear the image but keep the transforms
The reset() method will reset the transforms, set to startX and startY
and reset color, strokeColor and strokeWidth
These can also be called through resetup(), redraw() and restamp()
Still need to update these in GitHub and NPM... soon.

PATCHES
Added cache as a STYLE function - thanks Disco Drama for the inspiration

STYLE = {cache:true} // will cache all DisplayObjects - or cache
individual types, etc.
Fixed Dial press to jump to value when dial rotated
Remembered to update the version... this is new to ZIM Cat... takes some
getting used to.
Fixed ZIM Multi-asset Object - bug in multiple loads now fixed - thanks
Netanela
Also adjusted Multi-asset Object so if filename is the same as a
previous filename
this file gets the path added to its id.
Adjusted alp() to set any object with a tag property so the opacity of
the tag is set
This means a textArea.alp(.5) will actually set the alpha of the HMTL
tag as well
This does not work with textArea.alpha = .5 - only with the .alp() -
Thanks Cajoek for prompting.
Updated ZIM Sprite to work with multiple spritesheets for one Sprite -
Thanks Netanela for the suggestion.