ZIM Cat 03 patches - matches NPM ZIM 10.3.0
Fixed height variable in Dial - needed declaration when adding
useLabels. 1/18/21
Changed Layout dispose() to actually dispose and removed LayoutManager
enabled and disabled - see Layout above. 1/19/21
Adjusted Squiggle and Blob so they do not dispose events when changing
points - logic issue when updating memory - thanks Stefan Meyer for
find. 1/19/21
Added animate() waitedCall and waitedParams to sequence - these were not
being passed through - thanks Yan Wenge for the report.1/21/21
Fixed Label so backing goes beneath Label outline - we had set this to
work only with backgroundColor... 1/21/21
BREAK Turned back on letterSpacings when Label has a backgroundColor -
needed some way to handle kerning (still turned off for letterSpacing)
1/21/21
Adjusted Blob and Squiggle clone so like Shape, they clone the bounds.
These two do not always have perfect bounds - they take the default
bounds of the number of points.  Thanks Karel Rosseel for the report.
The approximateBounds() method can be used - but was being ignored when
cloned - for instance in an Emitter or Tile and so clone needs to clone
the new bounds.
Adjusted hitTestGrid() to properly take into account spaces - this came
from solving a bug in the Scrambler with spaces in a grid.
IMPROVEMENT Added a new type to hitTestGrid() called "open".
This allows spacing to be added and grid to return values even when
point is in spacing area
A grid needs to have the spacing added to the number of columns and
divided to get col or row sizes which had been done
But the position needs to have half the spacing NOT half the col or row
size - so this was adjusted.
The new setting has been applied to Tile for getting current tile under
mouse and Scrambler - which had a slight error of margin bug - now okay.
If you have a Scrambler in a grid with spacing then we recommend
updating to Cat 03.
Fixed Sprite run() with wait start, and loopWait with on return from
blurred window - thanks Netanela for the report.
The actual CreateJS Sprite needed to be explicitely paused in these
cases otherwise it would run on its own.
Changed noDrag() on draggable Panel to be noDrag(false) to keep cursors
on objects in panels - no need to recursively turn drags off
Added collapse and collapsed parameters to end of Panel to allow double
clicking on bar to hide and show panel content - leaving bar visible
Added collapsed getter setter property to Panel
BREAK adjusted tap() to be time in seconds by default rather than ms and
to default to 8 seconds for a tap.
Updated AlphaEffect to have proper examples - also made mask work with
Bitmap directly (rather than Bitmap.image)
and work with any ZIM DisplayObject - so GradientColor and RadialColor
will work for example - caches mask and uses its cacheCanvas as a mask
Added an AlphaEffect to the https://zimjs.com/cat/effects.html example -
it is the striped Pragma for a selected effect.
Added effects property to docs for Container, Shape, Bitmap, Sprite,
MoviClip