Patches to ZIM Cat 04 - various
The Button now activates rollover when pressed if the rollover has not already been activated.  This allows rollover color, backgroundColor, backings and icons to activate on mobile touch whereas before, they did not as there is no mouseover and mouseout events on mobile.  Along with this change we also now do not apply rollover colors if the mouse or press was already down outside the button and then the button is rolled on.  Thanks @omarpta and others in the past for this suggestion - we just figured out how to implement it - arg - go brain go!  Here is a simple example https://zimjs.com/test2/roll.html
The Frame now sets the canvas parent to position relative if no position is already set.  It turns out that that solves the Tag, TextArea and Loader positioning issues when scaling the canvas tag with CSS and retina is on (which is default).  Thanks @Stefan for the initial work here - it turns out that we think the solution is easier than expected.  Perhaps you can do some continued testing in that area - but it seems to work for us.  We remade the MEME app with ZIM Cat 04 and tested using the tag mode with the canvas scaled here: https://zimjs.com/meme/indexscaled.html
Also, @omarpta has made updates to ZIM Base to work with PHP 8.  Cheers!  A new ZIP is provided that also includes the SQL to make the tables used.