Variety of Patches to ZIM Cat 04
Possible some of these were in the last Commit - the patches at the end are the latest patches:

Fixed Pen so that it draws curves between each point rather than straight lines
The curves were there but were curving through a midpoint between points.
Well, this was making a straight line!
We needed to start at the previous midpoint and curve through the last point to the next midpoint
Thanks Ferudun Vural for pointing out that we still had straight lines when animated quickly.
This is an excellent change to Pen when reducing the damping (closer to 1)
and results in smoother lines, as was intended, when drawing at faster speeds.
Adjusted labels in Slider and Dial to be properly in the labels container.
Now slider.labels.cache() will work properly for instance (previously only the last label made would be cached)
Fixed List to ignore items without labels - little bug in ZIM Cat 04 when we added setting backgroundColor to clear if no label.
Fixed Window and List scroll with damping swipestop event was being overwritten by earlier scrollbar drag fixes.
Added bindings and GET url to TO and FROM report in Base - to more easily debug
And added bindings to report() along with existing data.
Updated ZIM Base with a new ZIP that gets rid of a PHP warning - thanks Amy Szczepanowski for the report
Updated ZIM NIO / Pizzazz 04 Paths https://zimjs.com/nio/paths.html to DIAMOND so TextArea selection works on Chrome
Fixed up a warning message in animate() on path about startPercent and percent not supported with drag
We were checking to warn if added but percent defaults to 100 so was warning all the time.
Added margin parameter to end of chop() to let pictures be chopped with extra margins - used for jigsaw puzzle for instance
Added a call2 parameter at the end tap() to call a function if a pressup but not a tap (or double tap)
Fixed Button color and rollColor direct parameter values being overwritten by label color and rollColor styles
Fixed up splitWords on Tips() after adjustment to Label parameters - thanks Ami Hanya for the report.
Patched Lazy Loading sound, mouseplus and fullscreen - see below.
Added fontListHeight and fontListViewNum to TextEditor - thanks Marva for the good idea.
Adjusted animate() sequenceCall parameter to be the object animating.  Thanks Yan Wenge.
Added button roll colors being applied on mousedown even if rollover does not occur - as in on mobile.
This allows the roll colors and icons to be applied as a mobile button is being pressed
If the user moves off while pressing down then the normal colors and icons apply, etc.
If the user taps then the color will go back to normal on mobile or stay as the roll color on non-mobile
Thanks omarpta for the request - and I think some others in the past - but we missed the solution.
Also changed the button rollover to not rollover if the mouse or press is already down outside the button then rolling over the button.
These changes were not tested with all formats of the Button like wait, toggle, icons and backings but they call the same roll functions so should be good.
Added a test to set canvas parent tag to position relative if no position is set.
This makes Tag() and TextArea(), etc. align properly when canvas tag is scaled.
Thanks Stefan Meyer for initial work on this - it turns out there was a more simple way to fix (found by accident!)
CheckBox rectangle and border rectangle now have box and box2 properties - was background property but never was working
ZIM chop has been updated to work without globals - we had left the namespace off loop, Bitmap and Tile.
The patch on ZIM Button has been patched to that.stage.frame rather than frame which does not necessarily exist.
Thanks Ashley Long for the report on these two issues.
We fixed the Synth Tone stop() docs to show the proper releaseTime parameter
And also fixed how stop() right on the tone to start works so it lets the sound fade in a little otherwise stop volume is default JS volume.
The duration parameter was calling stop and had the same problem so that is patched - thanks Yanwenge for the report!
Added pausedOnBlur parameter to Synth tone() to pause tone when window is reduced or goes to another tab - thanks Yanwenge for the request
This defaults to false at the moment but in Cat 05 we will switch it to true
Added paused(state) method to Synth tone() to mute and unmute
Fixed MovieClip in Cat 03 and Cat 04 - there was a typo in that rather than this.  Thanks Ferudun for the report.
Fixed Emitter dispose() when Frame is disposed - thanks Ashley Long for the report.
Adjusted List.slider to make using stepper work (typo in offset) and made it use the step provided
Adjusted List.colorPicker to make ColorPicker start with selected color - swatch did but not the default ColorPicker
Fixed ZIM Frame and CreateJS to scroll mobile canvas when allowDefault is true and yet avoid double events
This has been ongoing work by Ferudun Vural and Dan Zen and hopefully is fixed now.  Thanks Pan K for the reminder!
Patched CreateJS 1.3.3 but also launched this as 1.3.4 to update the NPM with new version.
A matching ZIM 11.4.1 on NPM has been provided.