Patches to ZIM Cat 04
Added call3 and call4 callback functions to tap() these activate on single tap when dbl is true
with call3 happening for any single tap and call4 happening on single tap only if the double tap fails
For instance if you want to do something on single tap but something else on double tap on the same object
then use obj.tap({dbl:true, call:doubleTapFunction, call4:singleTapFunction})
Thanks Xuntar for the use case suggestion - and the code addition for the Tap() to prevent double tap drag single tap call ;-)
Added local properties for Label font, bold, etc. when being set with properties... so clone and duplicate are correct - thanks Marva (again!)
Swapped the position of Pane and Panel in Docs so search on pane finds pane first and not panel.
Fixed hitTestCircleRect to work with a circle that is scaled.
Updated bounds of Label when labelWidth or labelHeight are set - thanks Vishwas for the report.
Fixed List when noScale is true so the selectedIndexPlusPosition goes to right place (only with equal or close to equal custom list sizes) - thanks Ami Hanya for report
Added animating and waiting properties to any object that animate() is operating on (also wiggle)
animating is the opposite of paused. animating is also true if wait, rewindWait, loopWait is happening
so also test the waiting property if you intend to see the animation occurring
Fixed ZIM ColorPicker which was using labelHeight and this had changed its functionality...
Fixed ZIM animate() with drag on path with speed set - was not animating after pauseAnimate(false);
Adjusted lazy loading of sound so that does not give an error if sound is not played right away