ZIM NFT 0 - Patches
See https://zimjs.com/updates.html for more - this lists all the patches so far to ZIM NFT 0:

Adjusted Label so it does not automatically set width to lineWidth if there is a lineHeight
This was causing scaling to break if a word was wider than the lineWidth - thanks Racheli Galan for the report.
Added borderColor to styles that can be set in List secondary hierarchy...
Added pulldownToggle parameter to end of List before style parameter
Set this to true to close the pulldown when selecting or pressing off pulldown.
Added "bloom", "expanded" and "contracted" events to List for accordian or pulldown functions.
These get an event object with item or items properties of the expended items - nothing for contracted.
Also added simple pulldown example to Docs for List.
Fixed ZIM List for pulldown with items not having labels - thanks yanwenge for the report.
Fixed accent on Dial when not in linear mode - was shifted 90 degrees as it was primarilay made for sound setting.
ZIM Container(a,b,c,d) has been updated to fix a glitch when providing only the a parameter
which was supposed to provide a square container with a as its width and height.
This was overwritten when we allowed a ZIM Boundary object to be the first parameter.
This has been fixed so if a has an x property it will assume a boundary object otherwise it will make a square.
We have also re-written the Container parameters in the Docs to be more clear on the options
Thank you Yingbing Jiang for the report.
Removed a notice about audio context being ready in CreateJS 1.3.4 - it was left in there from debugging.
Added a 10 ms delay to stage update on pane.show() so .mov() or .loc() can be used to reposition without a flash
Added reset to color and rotation for emitter particles when using the pool (default)
This makes color and rotation animation reset to the original as a new particles is created
Changed life parameter in Emitter to be a ZIM VEE value - thanks Chris Spolton for these suggestions
Also thanks Chris for pointing out clone was not cloning effects - we cloned the filters but needed to cache the cloned object for them to be seen
Added getMinMax() static method to ZIM Pick() to find min max range of ZIM VEE value (with numbers and not a function)
Pick.getMinMax(ZIM VEE value) - returns a {min:val, max:val} object
 	with the min and max values the ZIM VEE value can return
	or null, null if not numbers or the ZIM VEE is a function
	This will work for a range object {min, max}, an array of number [] or a series() of numbers
Added arrayMinMax(arr) function to code module to return {min, max} object for array
Fixed emitter for freeze on pause and for clearPool method
These were looping through the object rather than the object's particles
ever since ZIM 9.5.0 when particles were adjusted to be placed in a particles container.
Fixed Stepper to step numerically when min set - weird bug with thinking number was string despite min and max being set to Number()
And made currentValue report a Number when asked and Stepper is set to "number" (the default)
Patched Generator() to add output and outputType parameter to export to images (see section above)
Fixed ZIM TextInput to properly add() the placeholder to the window rather than addTo() when TextInput starts with text
Backed out of change to ZIM Synth that was added in ZIM Cat 04
to set volume of tone() if stop() is called right away on it - the fix was causing problems playing sound
so if you want to stop a tone right away, do not make the tone until you want to play it.
Please note - Yanwenge who reported the initial volume issue.
Added mouseChildren=false to Scrambler tile elements - thanks Netanela for the report
This allows Containers that are scrambled to work properly.
(could possibly make each element drag({all:true}) and leave mouseChildren on
in case interactivity inside tile item is needed - but there are complicated delays
If anyone needs this, let us know and we can probably adjust)
made literal version of series recursive
which was tricky because Pick.choose() was processing noPick coming back from recursive literal
adjusted ble() to update() Blob and Squiggle so blendMode gets applied immediately (bug since ZIM 6)
Added noSwipe() to Slider by default so they do not slide pages
Note that there were 77 patches to ZIM Cat 04 - all of which have improved ZIM NFT.