Patches to ZIM NFT 01 including GIF()
Added GIF() for reading animated gifs - needs gifler.js
Added an adjust parameter to ZIM SoundWave and a normalize parameter to calculate()
The normalize currently defaults to false but will default to true in the next version of ZIM
This allows SoundWave to calculate values between 0-1.
Adjusted rarity to store a Number() object if the key is a number and there is a payload
and to convert the key to a number primitive if key is number and there is no playload.
Adjusted loop() to convert a Number and String object to a number and string primitive before looping.
Warning... it is possible that a Number or String object might need to be cast as Number() or String() before being used.
For instance, in the FXHash features object, these must be primitives.
Fixed asset() for lazy loading to not intermitently load two images when the same id is used twice
For instance: asset("help.png").center(); asset("help.png").loc(100,100) should result in just one image at 100, 100.
It was intermittently, depending on cache showing two or one image.
Adjusted ZIM Swiper to use Frame "mouseupplus" event to capture mouse coming back from iFrame when no longer up. <strong>IMPROVEMENT</strong>
Adjusted physics.noDrag() to work with a list when there was not an initial draglist - for instance if a general physics.drag() was called.
Added ZIM DUO to physics.join()
These are patched in physics_2.1.js.
Added currentValue parameter to Stepper at end before style
Fixed animate() to work properly on calling multiple relative calls on same object when also moving object manually between <strong>IMPROVEMENT</strong>
reset the zimLastObj properties of the target - tested in general but keep watch for things like relative series, sequences, path, etc.
Fixed Scrambler when pressing up outside an iframe - used the Frame "mouseupplus" event <strong>IMPROVEMENT</strong>
There were 90 lines of patches for ZIM NFT 00 - these of course are all in ZIM NFT 01.
Added makeSyllable(lenght, firstVowel) to the CODE module - to make a syllable for generative words
Added CapitalizeFirst() function to CODE Module to capitalize the first letter of a string
Added a test to removeFrom() to removePhysics() if there is a removePhysics() method
so can removeFrom() without error on physics objects. Thanks Naythan Ragland for report.
Fixed glitch in expand() when using DUO - now works properly
Added a backingContainer to Dialog() so can cache the container and set alpha of backingContainer
otherwise the alpha will reveal the parts of the backing like the bubble and the tail - thanks Yanwenge
Added a margin to resetBounds() and made it accept DUO
so container.resetBounds({margin:10}) will recalculate bounds and add a margin of 10
but this would now be a static bounds rather than a dynamic bounds.
Made custom list work in EmojiPicker and added size parameter - thanks Karel Rosseel for the heads up.
Fixed Scrambler to provide getter setter order property to update the order dynamically.  Thanks Racheli Golan
Adjusted Docs for CamMotion - added visualizerObj - thanks Karel Rosseel for the find
Added cat zim.js versions back into CDN - had accidentally filtered out when we moved to zim_min.js for NFT
Fixed ZIM ISO example https://zimjs.com/iso/ to animate with time in seconds when capturing orb
Made ZIM Arrow() have the activate() method even if not being used with Pages - also fixed grey color which had () on it
Fixed Label with height to also use valign MIDDLE - thanks Ami Hanja.
Adjusted TextInput align right to properly align text entered in parameter at the right - thanks Racheli Golan for the report.
Added a rtl parameter at the end of Bitmap cache() and adjusted CreateJS 1.3.4 BitmapCache(), define(), updateCache() and update() with rtl
Thanks Marva for the solution