ZIM ZIM 02 added Pack and various updates
See https://zimjs.com/updates.html - this update adds Pack() and various patches such as:
Standardized on getPoints(), setPoints(), getSteps(), setSteps(), getOrder(), setOrder() for recording and retrieving
there are also related properties points, steps and order in Squiggle/Blob, Connectors and Pack respectively
Patched threejs OrbitControls v142 (the latest script version) - it was giving an error - thanks Karel Rosseel for the report
Added a removePoint(index) method to Squiggle and Blob - thanks Yanwenge for the suggestion - this just uses splice on the points property
Fixed ZIM MotionController - the flip:VERTICAL was backwards.
Adjusted ZIM Flare so cross colors can accept Gradients... previously it was only the whole shape that could.
See https://zimjs.com/editor/view/Z_GRQNU
Fixed ZIM Tabs so function STYLE does not go automatically to the Label - this was a bug as style needs a false rather than a null to deactivate
If STYLE for labels in Tabs (and Lists) is desired then styleLabels needs to be turned on.
Adjusted convertColor() so it handles RGB at 0 and 1 ranges and fixed animate() to not treat RGB string as relative values - thanks Ferudun Vural for note.
Adjusted the TextInput to place the placeholder at the label and match alignment.  Thanks Marva for the prompting.
Also made placeholder go away when setting the text with the text property or come back if text is set to "" or null.
Adjusted ZIM Distill to automatically add fastFrame which has a Stage as dependency if there is no Frame added.
This will allow Distill to work better with Adobe Animate and ZIM Shim.
Thanks to Максим Сельденков on YouTube for the report. NOTE - this is still being worked on...
Adjusted Frame loading of JSON to provide just the parsed JSON object rather than adding src, item and file properties like we do for images and sound
Added a "first" parameter to zet(selector, first) to return only the first result - so gets [querySelector()] rather than querySelectorAll()
Fixed a bug with ZIM animate() for negative relative animations using the dot property so "rotation.x":"-20" - this now works
Adjusted Swiper() to use the sensitivity to detect movement ending - thanks EZZ FLash for the report
we updated the threejs globe spinning to ZIM ZIM 02 with the patch https://codepen.io/zimjs/pen/qGPVqO
Adjusted dragBoundary() so that if null is passed (or no parameter value) then the drag boundary is cleared.
The boundary will exist until the next mousedown on the object so if wanting to clear as dragging set a boundary of the stage size or bigger.
Added a new example to ZIM drag() to show how to drag horizontally or vertically depending on start direction
Thanks Ferudun Vural, Pettis Brandon and Joseph Diefenbach for the help in Slack.
Fixed ZIM asset object with font {assets:[{font:"amboy", src:"amboy.woff"}, "cat_s.png"], path:"assets/"}
Which can now also be {assets:["amboy.woff", "cat_s.png"], path:"assets/"} - thanks Pettis Brandon
Added NONE as a constant - used in upcoming Pack() class
Fixed ZIM Window so the damp parameter works properly - List too...
Changed Docs to use F, S, W, H in examples rather than frame, stage, stageW and stageH
Fixed bug in ColorPicker "legacy" - this is now working - thanks Marva for the report.
Also cursor was not showing when ColorPicker spectrum was on second Frame - thanks Marva.